requirements.txt
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==45.0.3
faiss-cpu==1.11.0
Flask==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
numpy==2.2.6
packaging==25.0
pycparser==2.22
PyJWT==2.10.1
requests==2.32.3
urllib3==2.3.0
Werkzeug==3.1.3
